split large modules in cli to smaller modules
