+++ packages/cli/src/adapters/index.ts
@@ -1,5 +1,5 @@
 export * from "./injection";
 export * from "./logger.adapter";
-export * from "./node-workspace.adapter";
+export * from "./node-workspace";
 export * from "./prompt.adapter";
 export * from "./reporters";
