+++ packages/cli/src/adapters/node-workspace/index.ts
@@ -1,0 +1,2 @@
+export * from "./node-workspace";
+export * from "./workspace";
