+++ packages/cli/src/adapters/node-workspace/workspace.ts
@@ -1,0 +1,126 @@
+import fsp from "node:fs/promises";
+import path from "node:path";
+import { applyEdits, modify } from "jsonc-parser";
+
+export type Framework = "vite" | "bun" | "node";
+
+export type WorkspaceOptions = {
+	cwd?: string;
+};
+
+export class Workspace {
+	private cwd: string;
+
+	constructor(options: WorkspaceOptions = {}) {
+		this.cwd = options.cwd || process.cwd();
+	}
+
+	resolve(...paths: string[]) {
+		return path.resolve(this.cwd, ...paths);
+	}
+
+	async exists(filePath: string) {
+		try {
+			await fsp.access(this.resolve(filePath));
+			return true;
+		} catch {
+			return false;
+		}
+	}
+
+	async readFile(filePath: string) {
+		return fsp.readFile(this.resolve(filePath), "utf-8");
+	}
+
+	async writeFile(filePath: string, content: string) {
+		const fullPath = this.resolve(filePath);
+		await fsp.mkdir(path.dirname(fullPath), { recursive: true });
+		await fsp.writeFile(fullPath, content, "utf-8");
+	}
+
+	async detectFramework(): Promise<Framework> {
+		try {
+			const content = await this.readFile("package.json");
+			const pkg = JSON.parse(content);
+			const allDeps = { ...pkg.dependencies, ...pkg.devDependencies };
+
+			if (allDeps.vite) return "vite";
+			if (allDeps["@types/bun"] || allDeps.bun) return "bun";
+		} catch {
+			// ignore missing or invalid package.json
+		}
+
+		// Check for config files
+		if (
+			(await this.exists("vite.config.ts")) ||
+			(await this.exists("vite.config.js")) ||
+			(await this.exists("vite.config.mts")) ||
+			(await this.exists("vite.config.mjs"))
+		) {
+			return "vite";
+		}
+
+		// Check for bun runtime
+		if ("bun" in process.versions) return "bun";
+
+		return "node";
+	}
+
+	async setTsConfigProperty(propertyPath: string[], value: unknown) {
+		const tsConfigPath = await this.findTsConfig();
+		if (!tsConfigPath) return { status: "not_found" };
+
+		try {
+			const content = await this.readFile(tsConfigPath);
+			const edits = modify(content, propertyPath, value, {
+				formattingOptions: { insertSpaces: true, tabSize: 2 },
+			});
+			const updated = applyEdits(content, edits);
+
+			await this.writeFile(tsConfigPath, updated);
+			return { status: "updated", file: path.basename(tsConfigPath) };
+		} catch {
+			return { status: "error", file: path.basename(tsConfigPath) };
+		}
+	}
+
+	async findTsConfig(): Promise<string | null> {
+		const filenames = [
+			"tsconfig.app.json",
+			"tsconfig.json",
+			"tsconfig.base.json",
+			"tsconfig.node.json",
+		];
+		let currentDir = this.cwd;
+		let isRoot = false;
+
+		do {
+			isRoot = currentDir === path.parse(currentDir).root;
+			for (const file of filenames) {
+				const fullPath = path.join(currentDir, file);
+				try {
+					await fsp.access(fullPath);
+					return fullPath;
+				} catch {
+					// intentionally ignore missing file
+				}
+			}
+			currentDir = path.dirname(currentDir);
+		} while (!isRoot);
+
+		return null;
+	}
+
+	async findViteConfig(): Promise<string | null> {
+		const filenames = [
+			"vite.config.ts",
+			"vite.config.js",
+			"vite.config.mts",
+			"vite.config.mjs",
+		];
+		for (const file of filenames) {
+			if (await this.exists(file)) return this.resolve(file);
+		}
+		return null;
+	}
+}
