+++ packages/cli/src/adapters/reporters/text.reporter.ts
@@ -1,0 +1,60 @@
+import { spinner as clackSpinner, note, outro } from "@clack/prompts";
+import pc from "picocolors";
+import type { Reporter, Spinner } from "./types";
+
+export class TextReporter implements Reporter {
+	info(message: string) {
+		process.stdout.write(`${pc.blue(`ℹ ${message}`)}\n`);
+	}
+
+	warn(message: string) {
+		process.stderr.write(`${pc.yellow(`⚠ ${message}`)}\n`);
+	}
+
+	error(message: string) {
+		process.stderr.write(`${pc.red(`✘ ${message}`)}\n`);
+	}
+
+	success(message: string) {
+		process.stdout.write(`${pc.green(`✔ ${message}`)}\n`);
+	}
+
+	step(message: string) {
+		process.stdout.write(`○ ${message}\n`);
+	}
+
+	note(message: string, title?: string) {
+		note(message, title);
+	}
+
+	log(message: string) {
+		process.stdout.write(`${message}\n`);
+	}
+
+	spinner(): Spinner {
+		return clackSpinner();
+	}
+
+	json(data: unknown) {
+		process.stdout.write(`${JSON.stringify(data, null, 2)}\n`);
+	}
+
+	cancel(message: string) {
+		process.stderr.write(`${pc.red(`✘ ${message}`)}\n`);
+		process.exit(1);
+	}
+
+	fatal(message: string, error?: unknown) {
+		process.stderr.write(`${pc.red(`✘ ${message}`)}\n`);
+		if (error) {
+			process.stderr.write(
+				`${pc.red(error instanceof Error ? (error.stack ?? String(error)) : String(error))}\n`,
+			);
+		}
+		process.exit(1);
+	}
+
+	finish(message: string, _details?: Record<string, unknown>) {
+		outro(message);
+	}
+}
