--- packages/cli/src/cli/ui/prompts.ts
@@ -1,224 +1,0 @@
-import fs from "node:fs";
-import path from "node:path";
-import { cancel, confirm, group, isCancel, select, text } from "@clack/prompts";
-import pc from "picocolors";
-import { getEnvExampleKeys, type ProjectOptions } from "@/features/scaffold";
-import { code } from "./visuals";
-
-export async function runPromptWizard(
-	defaults?: {
-		framework?: ProjectOptions["framework"];
-	},
-	isYes = false,
-): Promise<ProjectOptions | null> {
-	const detectedKeys = await getEnvExampleKeys();
-
-	if (isYes) {
-		const framework = defaults?.framework || "node";
-		let envDtsHandling: ProjectOptions["envDtsHandling"];
-
-		if (framework === "vite" || framework === "bun") {
-			const typeFile = framework === "vite" ? "vite-env.d.ts" : "bun-env.d.ts";
-			const targetDir = path.dirname(
-				path.resolve(process.cwd(), "./src/env.ts"),
-			);
-			const typeFilePath = path.join(targetDir, typeFile);
-			envDtsHandling = fs.existsSync(typeFilePath) ? "append" : "overwrite";
-		}
-
-		return {
-			path: "./src/env.ts",
-			validator: "arktype",
-			framework,
-			language: "ts",
-			overwriteEnvSchemaFile: true,
-			envDtsHandling,
-			installTypeDefinitions: framework !== "node",
-			envKeys: detectedKeys || undefined,
-			installSkill: false,
-		} as ProjectOptions;
-	}
-
-	const result = await group(
-		{
-			overwriteEnvSchemaFile: async () => {
-				const defaultPath = "./src/env.ts";
-				if (fs.existsSync(path.resolve(process.cwd(), defaultPath))) {
-					const answer = await confirm({
-						message: pc.yellow(
-							`An existing ArkEnv configuration was found at ${code(defaultPath)}. Do you want to overwrite it?`,
-						),
-						initialValue: false,
-						active: "Yes (override my configuration)",
-						inactive: "No (abort)",
-					});
-					if (isCancel(answer)) {
-						cancel("Operation cancelled.");
-						process.exit(0);
-					}
-					if (!answer) {
-						cancel("Operation cancelled.");
-						process.exit(0);
-					}
-					return answer;
-				}
-				return true;
-			},
-			framework: () =>
-				select({
-					message: "Select your framework or runtime:",
-					initialValue: defaults?.framework,
-					options: [
-						{
-							value: "vite",
-							label: `Vite${defaults?.framework === "vite" ? " (Detected)" : ""}`,
-						},
-						{
-							value: "bun",
-							label: `Bun${defaults?.framework === "bun" ? " (Detected)" : ""}`,
-						},
-						{
-							value: "node",
-							label: `Node.js${defaults?.framework === "node" ? " (Detected)" : ""}`,
-						},
-					],
-				}),
-			useDefaultPath: () =>
-				confirm({
-					message: "Use default config path (./src/env.ts)?",
-					initialValue: true,
-					active: "Yes (Recommended)",
-					inactive: "No, let me customize it",
-				}),
-			path: async ({ results }) => {
-				if (!results.useDefaultPath) {
-					return text({
-						message: "Where should we create the ArkEnv config?",
-						placeholder: "./src/env.ts",
-						initialValue: "./src/env.ts",
-					});
-				}
-				return "./src/env.ts";
-			},
-			installTypeDefinitions: async ({ results }) => {
-				if (results.framework === "vite" || results.framework === "bun") {
-					const typeFile =
-						results.framework === "vite" ? "vite-env.d.ts" : "bun-env.d.ts";
-					const targetDir = path.dirname(
-						path.resolve(
-							process.cwd(),
-							(results.path as string) || "./src/env.ts",
-						),
-					);
-					const typeFilePath = path.join(targetDir, typeFile);
-
-					if (fs.existsSync(typeFilePath)) {
-						return true;
-					}
-
-					return confirm({
-						message: `Establish ${code(typeFile)} for typesafe environment variables?`,
-						initialValue: true,
-						active: "Yes (Recommended)",
-						inactive: "No",
-					});
-				}
-				return true;
-			},
-			envDtsHandling: async ({ results }) => {
-				if (!results.installTypeDefinitions) return "skip";
-				if (results.framework !== "vite" && results.framework !== "bun")
-					return "skip";
-
-				const typeFile =
-					results.framework === "vite" ? "vite-env.d.ts" : "bun-env.d.ts";
-				const targetDir = path.dirname(
-					path.resolve(
-						process.cwd(),
-						(results.path as string) || "./src/env.ts",
-					),
-				);
-				const typeFilePath = path.join(targetDir, typeFile);
-
-				if (fs.existsSync(typeFilePath)) {
-					const answer = await select({
-						message: `Found existing ${code(typeFile)}. How should we handle ArkEnv types?`,
-						options: [
-							{
-								value: "append",
-								label: "Append types safely (if needed)",
-								hint: "Recommended",
-							},
-							{
-								value: "overwrite",
-								label: "Overwrite entirely",
-								hint: "Destructive",
-							},
-							{ value: "skip", label: "Skip" },
-						],
-					});
-					if (isCancel(answer)) {
-						cancel("Operation cancelled.");
-						process.exit(0);
-					}
-					return answer as ProjectOptions["envDtsHandling"];
-				}
-				return "overwrite";
-			},
-			validator: () =>
-				select({
-					message: "Select your preferred validator library:",
-					options: [
-						{
-							value: "arktype",
-							label: "ArkType (Recommended)",
-							hint: "TypeScript's 1:1 validator, optimized from editor to runtime",
-						},
-						{
-							value: "zod",
-							label: "Zod",
-							hint: "TypeScript-first schema validation with static type inference",
-						},
-						{
-							value: "valibot",
-							label: "Valibot",
-							hint: "The modular and type safe schema library",
-						},
-					],
-				}),
-			useEnvExample: async () => {
-				if (detectedKeys) {
-					return confirm({
-						message: `Detected ${pc.cyan(".env.example")} with ${detectedKeys.length} keys. Use them for your schema?`,
-						active: "Yes (Recommended)",
-						initialValue: true,
-					});
-				}
-				return false;
-			},
-		},
-		{
-			onCancel: () => {
-				cancel("Operation cancelled.");
-				process.exit(0);
-			},
-		},
-	);
-
-	if (isCancel(result)) {
-		cancel("Operation cancelled.");
-		process.exit(0);
-	}
-
-	return {
-		path: (result.path as string) || "./src/env.ts",
-		validator: result.validator as ProjectOptions["validator"],
-		framework: result.framework as ProjectOptions["framework"],
-		language: "ts",
-		overwriteEnvSchemaFile: result.overwriteEnvSchemaFile as boolean,
-		envDtsHandling: result.envDtsHandling as ProjectOptions["envDtsHandling"],
-		installTypeDefinitions: result.installTypeDefinitions as boolean,
-		envKeys: result.useEnvExample ? (detectedKeys as string[]) : undefined,
-		installSkill: false, // Defaulting to false, will be overridden by orchestrator if needed
-	};
-}
